Introducing fleetId, improved fleet selection, css change
A unique FleetId is added to fleet objects and foreignFleets list which
make is more easily to look up a fleet in stationedFleets and
foreignFleets without to rely on the fleet name and or owner.

Display of ships in scanAreas should now work as expected.

Bugfixed the display and handling of foreign fleets on a planet where
another players fleet is handled as stationed.

Minor css change to apply a font for every html object to improve the UI
somehow. Lowered the text size of the log in addition.