Adding gfx, workaround for colonization
Colonization was breaking in reason of fleet overview listing cleanups,
those have been temporary commented out and will be fixed in the next
update.

Biggest feature change is the ongoing work on a small gfx engine, which
allows to incooperate images, finally going away from most of the flat
designs, but now dialogs and such are out of place..

Anyhow, feel free to clone to your desktop to give it a try!