Important gfx updates, working in Chrome/Firefox..
Doctype has been added so IE doesnt complain anymore.
But IE10 fails to run the code. I did not test with other versions, yet.

Chrome has to be started with "--allow-file-access-from-files" in order
to be able to load the local graphics, BUT playback has been fixed in
Chrome :)

Some small css change because Production List and Queue were broken.