Colonize bugfix, small css change
Colonization was completely broken due to a variable change. Clicking on
a fleet which is located in an uncolonized planet or an uncolonized
planet with a colony ship now properly shows a colonize button.

Also the fleet only becomes removed after colonizing the planet, not
already when clicking the colonize button.

Also the update in the fleetlisting is now updating the design counters
as expected.

Some css changes and small typo fix. Little changes in the UI.

Mainloop timer placed at 50, which should run smooth and stop lock
something.