Planet display changes
Planets which are located are now automatically shown by there
"displayName", this name can vary from the "name" given by a player.

This makes much more sense then having no name shown for a discovered
but unexplored planets.

Planets which are not yet explored are also shown in brackets in there
name.