Bugfixes, colonization cleanup, module gathering..
This update mostly is about bugfixes.

The planet dialog also now shows unknown planets and addionally for
known and unknown planets, shows the shortest travel distance.
This might come into play when a range for fleets is used, at the momemt
this always allows travelling no matter what.

Other player planets no correctly show or show not brackets when
explored.

Colonization with two colonization ships in one fleet caused an error,
this has been resolved. There still something to do with colonization
when a fleet is foreign and which fleet of the player will be shown for
drawing, if other fleets are present.

Module information can now be gathered by planet and fleet list
stationed/foreign fleet list of a planet.