Code cleanup, better unit display in planet detail screen
Better unit display and conversion of planet output up to 2 digits
precision for agriculture, production and research.

Some code cleanups of this dialog and also removal of string slice code
which was not required. Precision is now used instead.

Variable to store the interface for the planet detail view changed from
to generic "infoScreen" to "planetDetailScreen".