Colonization, fleet handling, planets overview..
First, some minor css changes, as some of the planets overview layout
was breaking due to introduced fixed precision in an earlier commit.

The latest changes where mostly about fleet handling and colonization of
planets, in particular when the fleets are foreign. Also colonization
now is completely possible if that planet is not colonized yet by
another player.

There also have been some minor cleanup of repeating code in the fleet
listing (strg/control + click on fleet) so that only fleets of the
player are shown and selected.

ForeignFleets now have an index (idx) pointing to the correct element
inside the player fleets.

Some minor cleanups of not anymore required floor and ceil's for planet
locations as all x and y values are now proper int values.