Ongoing work on fleets, fleet internals and colonization
All player only can see very limited of other players fleets - those are
handled as "foreignFleets". A fleet who's owner is not first on a planet
will always become a foreignFleet, but is internally for the player
himself shown as a regular fleet with different positioning around a
planet.
Also the players scan areas update as expected.
The current setup has 3 players set, switching is possible with Shift +
Tab-key.

Also colonization has undergone some work to make this setup possible
and it is possible to colonize a planet, even if there are other player
fleets in the planet orbit, which was buggy earlier.

Also the design creation has been simplified and  set equal for all
players, one explorer and one colony ship.
There is still a "error" that the fleet detail view does not show the
fleet ships when hovered if the fleet is foreign.