Better playerEnv logic for planets and fleets
Unknown planets are now completely unknown to the player, meaning that -
using playerEnv.planets - it is not longer possible to spy on this
planet, as this information only contains the very basic information to
display this planet to the player.
Also those planets, when a fleet is directed and the fleet/route
selected really shows the player.displayName, but not the other player
given name (only after discovery).

Also fleets do update properly, meaning if another player planet is in
scanRange, and a fleet is moved, another foreign fleet symbol is shown
to the player - showing that there are more fleets on this location,
this wasnt the case earlier.