Code change to discover fleets
The code change does provide that a fleet appears or disappears after a
fleet has landed on a planet and while its moving, this has to be
tested.
There should be some kind of detection of fleets during flight, or do
all sensors just work when located on a planet?
Right now the scanAreas do update with a small delay so foreign fleets
disappear on flight of the players fleet.