Change in galaxy generation, player planets, colonization
Now the galaxy is generated in a structured way but with some randomness
still to the planet positioniong. For now 56 planets.

Also the player planet is not exactly the first, second, third and so,
but a randomly chosen planet amongst all generated.

Last but not least, a bug in the colonization has been removed which
left a fleet still in the list of stationed fleets after colonization,
this is resolved.