Resolved duplicated fleets, foreign fleet display, planets
Work on more restricted planet object - those are other player planets,
but population and such is updated.. is ongoing.

Fleets where duplicated in the logic.fleets, leading to errors and
unrequired checks in the update of scan areas of the player.
Also foreign fleets which should have been shown, did not update
accordingly, this has been corrected.