Limited data for other player planets
Data has been more limited of what a player can actual see of a foreign
players planet. If a fleet is stationed, the planet fleets, stationed
and foreign are shown. But once there isnt any other fleet of the player
stationed, those information disappear as we lose the logical link.