Easier keyboard handling, switching player
Sometimes the pressing and holding "strg"/"control" got stuck on random
occasions, this should be now resolved.

Switching the player can now be done by pressing "p" which should be
much easier and not run into problems with the browser itself.

There is still a random bug, when trying to move the fleet on a foreign
planet, it seems this only occurs when the planet is discovered and not
in range when starting the game...