Introducing requestAnimationFrame
There is still a bug with planets, who have been disocvered and are
navigated to again, in reason of how the planet information are display
to the player.

This update only introduces the "requestAnimationFrame" method versus
the "setInterval" method which had been used to display animations and
calculate the mainloop.

There is still some kind of stutter and the reason for this kind of lack
could be the local maschine on which the game is tested on - it could be
just to slow.