Bugfixes - handling of logic versus envs
This commit finally bugfixes the planet handling and update on fleet
arrival.
Also the non-visible change to work with the logic.environments where
approperiate or "env" variable of the player instead of the logic one -
which should only be retrieved by the game and not by the player - has
been added/changed or removed where required.
There is no mirroring anymore going on which wasnt really required by
the logic to do, as we have all player related information in the
environments.