Planet overview panel working again, fixed planet names
The planet overview in the main menu bar had been broken, meaning after
selecting a filter - the list had been sorted properly, but hovering
over any of the items did not show any information as expected when the
filter was still active.
Now sorting by name, distance and size works correctly.
Also this feature now only makes use of the planet information the
player has gathered, avoiding the use of the logic environments, which
should not be available to the player directly.
Minor change in css to allow the new changed code to display the
activeFilter instead of the "active" one which was used to determine the
active selected planet.

Also a bug in the random planet naming has been fixed which lead to a
"undefinied" name while the value beeing out of range of the list.