Queued items removal bugfix
When the player screen had been reopened again after a planet change,
the items in the production queue did not have the proper classname
assigned to them and removing of items did fail once the current player
changed.

Now, after adding the queued item list to the list, all items can be
removed as desired.