Fixed detection of colonized planets in scanRange, change in galaxy generation
If a planet is not colonized but does become colonized by another
player, the planet display for the player becomes updated - but only if
this planet is in scanRange of the player, otherwise the information
become updated once a fleet rediscover this happening.

Changed genGalaxy slightly to allow closer planets, try ding to avoid
the case that a player has no planets near by on start,

Readded logic.scanAreas as those are required by the logic to validate.
Changed player to amount by default to 4.