Work on IE10 comaptiblity, AI feauture...
Added function to determine if a fleet is stationed/local or on the
foreignList.

Set destination now uses addional player as information to get the
correct environment, this could be replaced for the current env of the
player later on.

Added hasModule function which checks if a module is incooperated in a
design.

Added trimming to planet colonization naming, if the string is empty,
the default name is used as settled earlier.

Added some AI functionality to begin with. AI does scout using
non-colony fleets and colonize if colony ships are available.

Added playerTypes ("human", "ai") to logic.